pyproject.toml
[tool.poetry]
name = "marcosanguineti-website"
version = "0.1.0"
description = "⛵️ Streamlit based Personal Website"
authors = ["GitMarco27 <marco.sanguineti.info@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.37.1"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
